+++ lib/screens/splash_screen.dart
@@ -24,13 +24,6 @@ class _SplashScreenState extends State<SplashScreen> {
     });
   }
 
-  // _navigateToHome() async {
-  //   await Future.delayed(const Duration(seconds: 3), () {
-  //     Navigator.pushReplacement(
-  //         context, MaterialPageRoute(builder: (context) => const Onboarding()));
-  //   });
-  // }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
